genql-: moved types
