+++ genql-cli/src/render/typeMap/scalarType.ts
@@ -1,5 +1,5 @@
 import { GraphQLEnumType, GraphQLScalarType } from 'graphql'
 import { RenderContext } from '../common/RenderContext'
-import { Type } from './renderTypeMap'
+import { Type } from 'genql-runtime/dist/types'
 
 export const scalarType = (type: GraphQLScalarType | GraphQLEnumType, _: RenderContext): Type => ({ name: type.name })
