+++ genql-cli/src/render/typeMap/unionType.ts
@@ -1,6 +1,6 @@
 import { GraphQLUnionType } from 'graphql'
 import { RenderContext } from '../common/RenderContext'
-import { FieldMap, Type } from './renderTypeMap'
+import { FieldMap, Type } from 'genql-runtime/dist/types'
 
 export const unionType = (type: GraphQLUnionType, _: RenderContext) => {
   const typeObj: Type & { fields: FieldMap } = {
