+++ genql-runtime/src/client/getFieldFromPath.ts
@@ -1,4 +1,4 @@
-import { LinkedField, LinkedType } from './linkTypeMap'
+import { LinkedField, LinkedType } from '../types'
 import startsWith from 'lodash.startswith'
 
 export const getFieldFromPath = ( // types.json is only needed for getFieldFormPath, can be removed? uses .fields and .type
